Bugfix: Fixes NaNs appearing in P & Sigma calculation.
P & Sigma were not being computed correctly for some C-ARMA models
because the integrand was underflowing for large integration variable.
To fix this, double maxT and double InitStepSize have been introduced
as C-ARMA class attributes with (temporary) default hardcoded values
maxT = 1.0e300 and InitStepSize = 1.0e-12. These values will be exposed
to the Python interface and set in the ConfigFile in the future.