Enable usage of GNU C++ Compiler
It is desirable to be able to use the GNU C++ Compiler instead of the
Intel C++ Compiler since the GNU Compiler is available free of charge.
This commit modifies the setup.py to check for the Intel C++ Compiler
but default to the GNU C++ Compiler if the Intel Compiler is
unavailable.

NB: MKL is still a requirement.